Fixes for decoding lossless_pfm conformance test image.
Uses wrapping_add and wrapping_sub for i32 operations.

Disable error for reading more than 29 bits for a symbol.

+++ jxl/src/frame/modular/transforms/palette.rs
@@ -221,7 +221,7 @@ pub fn do_palette_step_general(
                         palette_entry
                     };
                     out.data.as_rect_mut().row(y)[x] = val;
-                    wp_state.update_errors(val as i64, (x, y), w);
+                    wp_state.update_errors(val, (x, y), w);
                 }
             }
         }
